Refactor and enhance DHCP support in axnet-ng module
- Updated the EthernetDevice structure to support optional IPv4 addresses.
- Modified the init_network function to handle static and DHCP configurations.
- Implemented DHCP state management in the Service struct, allowing for dynamic IP address assignment.
- Added DHCP polling logic to the Service to handle address configuration and deconfiguration events.
- Enhanced routing table management to accommodate dynamic IPv4 configurations.
- Updated Cargo.toml files to enable DHCP-related features in the axnet-ng and axnet modules.
- Introduced a new test configuration for DHCP functionality in the qemu-x86_64 environment.
- Refactored various functions to improve code clarity and maintainability.

+++ components/starry-smoltcp/examples/loopback.rs
@@ -45,24 +45,22 @@ mod mock {
 mod mock {
     use std::sync::{
         Arc,
-        atomic::{AtomicUsize, Ordering},
+        atomic::{AtomicU64, Ordering},
     };
 
     use smoltcp::time::{Duration, Instant};
 
-    // should be AtomicU64 but that's unstable
     #[derive(Debug, Clone)]
     #[cfg_attr(feature = "defmt", derive(defmt::Format))]
-    pub struct Clock(Arc<AtomicUsize>);
+    pub struct Clock(Arc<AtomicU64>);
 
     impl Clock {
         pub fn new() -> Clock {
-            Clock(Arc::new(AtomicUsize::new(0)))
+            Clock(Arc::new(AtomicU64::new(0)))
         }
 
         pub fn advance(&self, duration: Duration) {
-            self.0
-                .fetch_add(duration.total_millis() as usize, Ordering::SeqCst);
+            self.0.fetch_add(duration.total_millis(), Ordering::SeqCst);
         }
 
         pub fn elapsed(&self) -> Instant {
